23-multi-user-rest-api-start/requirements.txt
asgiref==3.11.0
Django==5.2
pillow==12.1.0
sqlparse==0.5.5
tzdata==2025.3
djangorestframework==3.17.0